+++ examples/timer/README.md
@@ -1,0 +1,17 @@
+# Timer Example
+
+This is a technical demonstration for how to use timeouts and intervals.
+
+## Concepts
+
+The example mainly demonstrates the use of [`TimeoutService`] and [`IntervalService`]
+but also makes use of some more advanced [`ConsoleService`] features.
+
+## Improvements
+
+- Use CSS to improve the presentation of the example
+- Apply the concept to something more fun than just a dry technical demonstration
+
+[`timeoutservice`]: https://docs.rs/yew/latest/yew/services/struct.TimeoutService.html
+[`intervalservice`]: https://docs.rs/yew/latest/yew/services/struct.IntervalService.html
+[`consoleservice`]: https://docs.rs/yew/latest/yew/services/struct.ConsoleService.html
